+++ tests/unit/spec/src/components/modal/lib/providers/transition.test.js
@@ -1,7 +1,7 @@
 /** @jsx h */
 import { h } from 'preact';
 import { renderHook, act } from '@testing-library/preact-hooks';
-import { useXProps } from 'src/components/lib/providers/xprops';
+import { useXProps } from 'src/components/modal/lib/providers/xprops';
 import { TransitionStateProvider, useTransitionState, STATUS } from 'src/components/modal/lib/providers/transition';
 
 const defaultXProps = {
@@ -12,7 +12,7 @@ const defaultXProps = {
     onShow: jest.fn()
 };
 
-jest.mock('src/components/lib/providers/xprops');
+jest.mock('src/components/modal/lib/providers/xprops');
 
 describe('transition', () => {
     useXProps.mockReturnValue(defaultXProps);